Fix for 'ProductAPI.getProduct' for Sandbox.
On sandbox, it appears that the '/products' endpoint rotates values. If you try to get 'BTC-USD' while it is out of the rotation, it will not return a product. However, getting the '/products/productId' endpoint retrieves the requested data regardless if it exists inside the '/products' endpoint at that time.

This can be tested by refreshing the endpoint directly to witness the rotation of products. 
Sandbox Products: https://api-public.sandbox.pro.coinbase.com/products