Move test for undefined socket closer to usage
We found that if sendMessage() was called but not awaited on, the
code execution would get to the awaited signRequest() call and then
immediately return execution to caller of sendMessage() before
signRequest() returns. If the caller then calls disconnect(), when
signRequest() returns, the socket would be undefined, even though
our test for undefined above would find it existing.